feat(api): move all update_.+_in_db functions to database module
