+++ api/src/database/resources/flavor.rs
@@ -2,12 +2,16 @@
 
 use anyhow::Context;
 use avina_wire::resources::{
-    Flavor, FlavorCreateData, FlavorDetailed, FlavorGroupMinimal, FlavorMinimal,
+    Flavor, FlavorCreateData, FlavorDetailed, FlavorGroupMinimal,
+    FlavorMinimal, FlavorModifyData,
 };
 use sqlx::{Executor, FromRow, MySql, Transaction};
 
-use crate::error::{
-    MinimalApiError, NotFoundOrUnexpectedApiError, UnexpectedOnlyError,
+use crate::{
+    database::resources::flavor_group::select_flavor_group_name_from_db,
+    error::{
+        MinimalApiError, NotFoundOrUnexpectedApiError, UnexpectedOnlyError,
+    },
 };
 
 /// Select the name of the flavor with the given ID from the database, or return [None].
@@ -509,4 +513,49 @@ pub async fn delete_flavor_from_db(
         ));
     }
     Ok(())
+}
+
+#[tracing::instrument(name = "update_flavor_in_db", skip(data, transaction))]
+pub async fn update_flavor_in_db(
+    transaction: &mut Transaction<'_, MySql>,
+    data: &FlavorModifyData,
+) -> Result<Flavor, NotFoundOrUnexpectedApiError> {
+    let row = select_flavor_from_db(transaction, data.id as u64).await?;
+    let name = data.name.clone().unwrap_or(row.name);
+    let openstack_id = data.openstack_id.clone().unwrap_or(row.openstack_id);
+    let weight = data.weight.unwrap_or(row.weight);
+    let group = data.group.unwrap_or(row.group);
+    let query = sqlx::query!(
+        r#"
+        UPDATE resources_flavor
+        SET name = ?, openstack_id = ?, weight = ?, group_id = ?
+        WHERE id = ?
+        "#,
+        name,
+        openstack_id,
+        weight,
+        group,
+        data.id,
+    );
+    transaction
+        .execute(query)
+        .await
+        .context("Failed to execute update query")?;
+    let group_name = if let Some(group_id) = group {
+        Some(
+            select_flavor_group_name_from_db(transaction, group_id as u64)
+                .await?,
+        )
+    } else {
+        None
+    };
+    let project = Flavor {
+        id: data.id,
+        name,
+        openstack_id,
+        weight,
+        group,
+        group_name,
+    };
+    Ok(project)
 }
