+++ api/src/routes/resources/flavor_group/modify.rs
@@ -3,17 +3,14 @@ use actix_web::{
     web::{Data, Json, Path, ReqData},
 };
 use anyhow::Context;
-use avina_wire::{
-    resources::{FlavorGroup, FlavorGroupModifyData},
-    user::User,
-};
-use sqlx::{Executor, MySql, MySqlPool, Transaction};
+use avina_wire::{resources::FlavorGroupModifyData, user::User};
+use sqlx::MySqlPool;
 
 use super::FlavorGroupIdParam;
 use crate::{
     authorization::require_admin_user,
-    database::resources::flavor_group::select_flavor_group_from_db,
-    error::{NotFoundOrUnexpectedApiError, OptionApiError},
+    database::resources::flavor_group::update_flavor_group_in_db,
+    error::OptionApiError,
 };
 
 #[tracing::instrument(name = "flavor_group_modify")]
@@ -43,38 +40,4 @@ pub async fn flavor_group_modify(
     Ok(HttpResponse::Ok()
         .content_type("application/json")
         .json(flavor_group))
-}
-
-#[tracing::instrument(
-    name = "update_flavor_group_in_db",
-    skip(data, transaction)
-)]
-pub async fn update_flavor_group_in_db(
-    transaction: &mut Transaction<'_, MySql>,
-    data: &FlavorGroupModifyData,
-) -> Result<FlavorGroup, NotFoundOrUnexpectedApiError> {
-    let row = select_flavor_group_from_db(transaction, data.id as u64).await?;
-    let name = data.name.clone().unwrap_or(row.name);
-    let project = data.project.unwrap_or(row.project);
-    let query = sqlx::query!(
-        r#"
-        UPDATE resources_flavorgroup
-        SET name = ?, project_id = ?
-        WHERE id = ?
-        "#,
-        name,
-        project,
-        data.id,
-    );
-    transaction
-        .execute(query)
-        .await
-        .context("Failed to execute update query")?;
-    let project = FlavorGroup {
-        id: data.id,
-        name,
-        project,
-        flavors: row.flavors,
-    };
-    Ok(project)
 }
